Use a React Portal to display hints in header
After some back and forth, thinking about pulling up much of the state
that is used to determine which keys to show/fade in the hints, I
decided to go with a portal. This prevents us from having to 1) cram a
lot more state-management into the top-level <App> and 2) passing
through all the state values and setters.

The keyboard hints also are most closely tied to the <List>, so
co-locating them seems appropriate.

Also, the hints might be projected elsewhere some time, e.g. into a
proper tooltip bubble or similar, in which case, the portal could still
be useful.